[REFACTOR] GetFilteredHomeworkService에서 GetHomeworkContentService를 분리
